Unwrap plug errors with handle_errors
Let `Plug.ErrorHandler` do most of the work for us :)

This gives us unwrapped exceptions for free. The only downside that I
can see at this point is that we've lost the stack trace in the console
output. I'm going to try and sort that out before I land the PR.